Merge stage 2
Conflicts:
	build/mpsExecutionPlugin.xml
	core/findUsages-runtime/models/findalgorithm/finders.mps
	core/generator/test/jetbrains/mps/generator/test/GenerationTestBase.java
	core/tool/builder/models/jetbrains/mps/tool/builder.mps
	core/tool/builder/models/jetbrains/mps/tool/builder/make.mps
	core/tool/builder/models/jetbrains/mps/tool/builder/paths.mps
	core/tool/builder/models/jetbrains/mps/tool/builder/unittest.mps
	core/tool/builder/models/jetbrains/mps/tool/builder/util.mps
	core/tool/builder/models/jetbrains/mps/tool/common/util.mps
	core/tool/builder/models/jetbrains/mps/tool/environment.mps
	languages/baseLanguage/baseLanguage/solutions/jetbrains.mps.baseLanguage.test/jetbrains/mps/editorTest@tests.mps
	languages/baseLanguage/baseLanguage/solutions/jetbrains.mps.baseLanguage.unittest/jetbrains/mps/baseLanguage/test@tests.mps
	languages/baseLanguage/baseLanguage/source_gen/jetbrains/mps/baseLanguage/typesystem/TypesystemDescriptor.java
	languages/baseLanguage/closures/solutions/closures.test/jetbrains/mps/closures/test/helper@tests.mps
	languages/baseLanguage/collections/solutions/internalCollections.test/jetbrains/jetbrains/mps/internalCollections/test/typechecking@tests.mps
	languages/languageDesign/stubs/languageModels/constraints.mps
	languages/languageDesign/stubs/languageModels/editor.mps
	plugins/execution-configurations/languages/unitTests/runtime/jetbrains/mps/baseLanguage/unitTest/execution.mps
	plugins/execution-configurations/languages/unitTests/runtime/jetbrains/mps/baseLanguage/unitTest/execution/settings.mps
	plugins/execution-configurations/languages/unitTests/runtime/jetbrains/mps/baseLanguage/unitTest/execution/tool.mps
	plugins/execution-configurations/languages/unitTests/startup/models/jetbrains/mps/baseLanguage/unitTest/execution/server.mps
	plugins/execution-languages/languages/jetbrains.mps.execution.impl/jetbrains/mps/execution/impl/configurations.mps
	plugins/mps-build/languages/build.mps.tests/generator/template/jetbrains/mps/build/mps/tests/generator/template/main@generator.mps
	plugins/mps-build/languages/solutions/jetbrains.mps.build.mps.runtime/models/jetbrains/mps/build/mps/runtime/anttask.mps
	plugins/mps-build/solutions/mpsBuild/source_gen/jetbrains/mps/ide/build/mpsExecutionPlugin.xml
	plugins/mps-testing/languages/languageDesign/test/languageModels/scripts.mps
	plugins/mps-testing/languages/languageDesign/test/solutions/jetbrains.mps.lang.test.runtime/jetbrains/mps/lang/test/runtime.mps
	plugins/mps-testing/languages/languageDesign/test/solutions/jetbrains.mps.lang.test.runtime/source_gen/jetbrains/mps/lang/test/runtime/TransformationTestRunner.java
	plugins/mps-testing/languages/languageDesign/test/solutions/sandbox/models/jetbrains/mps/lang/test/sandbox.mps
	plugins/mpsjava/basePlatform/models/jetbrains/mps/ide/java/parser.mps
	plugins/mpsjava/jetbrains.mps.ide.java.testMaterial/java/testMaterial4.mps
	plugins/mpsjava/jetbrains.mps.ide.java.testMaterial/jetbrains/mps/ide/java/testMaterial/testModel1.mps
	plugins/mpsjava/jetbrains.mps.ide.java.testMaterial/jetbrains/mps/ide/java/testMaterial/testModel1/sub.mps
	plugins/mpsjava/jetbrains.mps.ide.java.testMaterial/jetbrains/mps/ide/java/testMaterial/testModel2.mps
	plugins/mpsjava/jetbrains.mps.ide.java.testMaterial/jetbrains/mps/ide/java/testMaterial/testModel2/sub.mps
	plugins/mpsjava/jetbrains.mps.ide.java.testMaterial/jetbrains/mps/ide/java/testMaterial/testModel3.mps
	plugins/mpsjava/jetbrains.mps.ide.java.testMaterial/jetbrains/mps/ide/java/testMaterial/testModel3/sub.mps
	plugins/mpsjava/jetbrains.mps.ide.java.tests/models/jetbrains/mps/ide/java/tests/madeUpCases@tests.mps
	plugins/mpsjava/jetbrains.mps.ide.java.tests/models/jetbrains/mps/ide/java/tests/utility.mps
	plugins/vcs/models/jetbrains/mps/vcs/diff/ui/merge@tests.mps
	plugins/vcs/models/jetbrains/mps/vcs/suspicious.mps
	plugins/vcs/models/jetbrains/mps/vcs@tests.mps
	testbench/modules/testActions/models/jetbrains/mps/testActions/test@tests.mps
	testbench/modules/testActions/models/jetbrains/mps/testActions/util.mps
	testbench/solution/jetbrains/mps.mps
	testbench/solution/jetbrains/mps/testbench.mps
	testbench/solution/jetbrains/mps/testbench/junit.mps
	testbench/solution/jetbrains/mps/testbench/junit/runners.mps
	testbench/solution/jetbrains/mps/testbench/junit/suites.mps
	testbench/solution/jetbrains/mps/testbench/util.mps
	testbench/solutions/jetbrains/mps/testbench.mps
	testbench/testbench.iml
	tools/gensources/gensources.iml
	workbench/mps-debugger/models/breakpoints.mps
	workbench/mps-debugger/models/core.mps
	workbench/mps-ui/models/jetbrains/mps/ide/ui/dialogs/properties.mps
	workbench/mps-ui/models/jetbrains/mps/ide/ui/dialogs/properties/renderers.mps
	workbench/mps-workbench/models/depanalyzer.mps
	workbench/mps-workbench/models/dialogs/project/components/parts/lists.mps
	workbench/mps-workbench/models/dialogs/project/components/parts/renderers.mps
	workbench/mps-workbench/models/dialogs/project/creation.mps
	workbench/mps-workbench/models/newSolutionDialog.mps
	workbench/mps-workbench/source_gen/jetbrains/mps/ide/actions/InternalActionsUtils.java

+++ languages/baseLanguage/baseLanguage/source_gen/jetbrains/mps/baseLanguage/behavior/Classifier_Behavior.java
@@ -174,7 +174,7 @@ public class Classifier_Behavior {
       containers = newContainers;
     }
 
-    if ((int) ListSequence.fromList(containers).count() == 1) {
+    if (ListSequence.fromList(containers).count() == 1) {
       return SPropertyOperations.getString(thisNode, "name");
     }
