+++ languages/baseLanguage/baseLanguage/source_gen/jetbrains/mps/baseLanguage/scripts/MembersMigrationUtil.java
@@ -44,7 +44,7 @@ public class MembersMigrationUtil {
         return linkName.equals(SPropertyOperations.getString(it, "role"));
       }
     }));
-    return ((int) ListSequence.fromList(links).count() == 1 ? ListSequence.fromList(links).getElement(0) : null);
+    return (ListSequence.fromList(links).count() == 1 ? ListSequence.fromList(links).getElement(0) : null);
   }
 
   private static SNode findNewBehaviorMethod(final String methodName) {
@@ -61,7 +61,7 @@ public class MembersMigrationUtil {
         return methodName.equals(SPropertyOperations.getString(it, "name"));
       }
     }));
-    return ((int) ListSequence.fromList(methods).count() == 1 ? ListSequence.fromList(methods).getElement(0) : null);
+    return (ListSequence.fromList(methods).count() == 1 ? ListSequence.fromList(methods).getElement(0) : null);
   }
 
   private static Iterable<SNode> classifierConcepts() {
